Refactor Gemini provider helper modules
